+++ src/sdk-gen/managers/index.ts
@@ -77,6 +77,8 @@ export * from './enterpriseConfigurations';
 export * from './hubs';
 export * from './hubCollaborations';
 export * from './hubItems';
+export * from './hubDocument';
 export * from './shieldLists';
 export * from './archives';
 export * from './externalUsers';
+export * from './automateWorkflows';