+++ CHANGELOG.md
@@ -2,6 +2,50 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [4.9.0](https://github.com/box/box-node-sdk/compare/v4.8.0...v4.9.0) (2026-05-06)
+
+
+### ⚠ BREAKING CHANGES
+
+* **boxsdkgen:** Mark `id` field of `AIAgentReference` required (box/box-openapi[#1629](https://github.com/box/box-node-sdk/issues/1629)) (box/box-openapi[#595](https://github.com/box/box-node-sdk/issues/595)) ([#1439](https://github.com/box/box-node-sdk/issues/1439)) ([3ab5519](https://github.com/box/box-node-sdk/commit/3ab5519c91695b1b3f42092655f57baa957ae0ef))
+
+### Bug Fixes
+
+* **boxsdkgen:** Fix metadata template schema (box/box-openapi[#596](https://github.com/box/box-node-sdk/issues/596)) ([#1442](https://github.com/box/box-node-sdk/issues/1442)) ([11aa977](https://github.com/box/box-node-sdk/commit/11aa97767a4653449b43da7a0397dc40abf39e97))
+* **boxsdkgen:** Mark `id` field of `AIAgentReference` required (box/box-openapi[#1629](https://github.com/box/box-node-sdk/issues/1629)) (box/box-openapi[#595](https://github.com/box/box-node-sdk/issues/595)) ([#1439](https://github.com/box/box-node-sdk/issues/1439)) ([3ab5519](https://github.com/box/box-node-sdk/commit/3ab5519c91695b1b3f42092655f57baa957ae0ef))
+
+## [4.8.0](https://github.com/box/box-node-sdk/compare/v4.7.0...v4.8.0) (2026-04-30)
+
+
+### New Features and Enhancements
+
+* **boxsdkgen:** Support Automate API (box/box-openapi[#594](https://github.com/box/box-node-sdk/issues/594)) ([#1432](https://github.com/box/box-node-sdk/issues/1432)) ([010fcdc](https://github.com/box/box-node-sdk/commit/010fcdc947003675403b4acc9d9a1485ee9c4bac))
+
+## [4.7.0](https://github.com/box/box-node-sdk/compare/v4.6.0...v4.7.0) (2026-04-22)
+
+
+### Bug Fixes
+
+* **boxsdkgen:** Update `retentionLength` field to accept both string and number (box/box-codegen[#932](https://github.com/box/box-node-sdk/issues/932)) ([#1396](https://github.com/box/box-node-sdk/issues/1396)) ([556f4a6](https://github.com/box/box-node-sdk/commit/556f4a6244e8a754c9c332da4b04dd548a57c631))
+
+
+### New Features and Enhancements
+
+* **boxsdkgen:** Add `max_extension_length` field to Retention Policies (box/box-openapi[#593](https://github.com/box/box-node-sdk/issues/593)) ([#1412](https://github.com/box/box-node-sdk/issues/1412)) ([a9a3fce](https://github.com/box/box-node-sdk/commit/a9a3fce3432bf6e89e759123b52f43b2e174360d))
+
+## [4.6.0](https://github.com/box/box-node-sdk/compare/v4.5.0...v4.6.0) (2026-04-01)
+
+
+### Bug Fixes
+
+* **boxsdkgen:** Correct search content type 'tags' to match API (box/box-openapi[#591](https://github.com/box/box-node-sdk/issues/591)) ([#1383](https://github.com/box/box-node-sdk/issues/1383)) ([d6bca72](https://github.com/box/box-node-sdk/commit/d6bca725910f2539bfb8ac984295e0c19b610191))
+
+
+### New Features and Enhancements
+
+* **boxsdkgen:** Add `reference` to `AI` and `parent_id` to `HubsItems` (box/box-openapi[#588](https://github.com/box/box-node-sdk/issues/588)) ([#1371](https://github.com/box/box-node-sdk/issues/1371)) ([5c299f0](https://github.com/box/box-node-sdk/commit/5c299f09268c546528c1748ea6e0774312a3f9f0))
+* **boxsdkgen:** Support Hub Document API (box/box-codegen[#930](https://github.com/box/box-node-sdk/issues/930)) ([#1385](https://github.com/box/box-node-sdk/issues/1385)) ([306c96c](https://github.com/box/box-node-sdk/commit/306c96cb988f0091fb146e1a4b15c12284f6d572))
+
 ## [4.5.0](https://github.com/box/box-node-sdk/compare/v4.4.0...v4.5.0) (2026-03-18)